chore: move older code into v1 gateway file
